* New setting exists: Keep screen on while open * Most of the timer system works.
Make-so:
* Dynamic UI gets cleared when pressing the stop button.
* RefreshDynamicUI does what it's supposed to.
* Countdown lock system (on the time-left/time-over label/button) works.

+++ Main/VDF/VDFLoader.cs
@@ -124,7 +124,9 @@ public static class VDFLoader
 			else if (firstNonMetadataToken.text.Contains(".") || firstNonMetadataToken.text.Contains("e"))
 				node.primitiveValue = double.Parse(firstNonMetadataToken.text);
 			else
-				node.primitiveValue = int.Parse(firstNonMetadataToken.text);
+				// custom replaced
+				//node.primitiveValue = int.Parse(firstNonMetadataToken.text);
+				node.primitiveValue = long.Parse(firstNonMetadataToken.text);
 		else if (firstNonMetadataToken.type == VDFTokenType.String)
 			node.primitiveValue = firstNonMetadataToken.text;
